+++ frontend/web/admin/dashboardDescription.md
@@ -1,0 +1,155 @@
+# Description des Dashboards par Type d'Administrateur
+
+Ce document détaille la structure et les éléments des dashboards pour chaque type d'administrateur dans le système Alpha Laundry Management.
+
+## 1. Master Super Admin Dashboard
+
+### Sidebar Elements
+- 📊 Vue d'ensemble
+- 👥 Gestion des Administrateurs
+- 🏢 Gestion des Entreprises
+- 📈 Statistiques Globales
+- ⚙️ Configuration Système
+- 🔐 Gestion des Permissions
+- 💰 Finance Globale
+
+### Layouts Détaillés
+
+#### Vue d'ensemble
+- KPIs principaux (nombre total de commandes, revenus, utilisateurs actifs)
+- Graphiques de performance globale
+- Alertes système
+- Activités récentes de tous les administrateurs
+
+#### Gestion des Administrateurs
+- Liste complète des administrateurs
+- Création/Modification/Suppression d'administrateurs
+- Attribution des rôles et permissions
+- Historique des actions administratives
+
+#### Gestion des Entreprises
+- Vue d'ensemble des entreprises affiliées
+- Statistiques par entreprise
+- Gestion des contrats et licences
+- Configuration des services par entreprise
+
+## 2. Super Admin Dashboard
+
+### Sidebar Elements
+- 📊 Tableau de Bord
+- 👥 Gestion du Personnel
+- 📦 Gestion des Commandes
+- 👤 Gestion Clients
+- 📈 Rapports
+- ⚙️ Paramètres
+
+### Layouts Détaillés
+
+#### Tableau de Bord
+- KPIs de l'entreprise
+- Graphiques de performance
+- Liste des tâches urgentes
+- Notifications système
+
+#### Gestion du Personnel
+- Liste des employés
+- Planning des livraisons
+- Performance des employés
+- Attribution des tâches
+
+## 3. Secretary Dashboard
+
+### Sidebar Elements
+- 📊 Aperçu Quotidien
+- 📝 Gestion des Commandes
+- 👤 Clients
+- 📋 Factures
+- 📅 Planning
+
+### Layouts Détaillés
+
+#### Aperçu Quotidien
+- Commandes du jour
+- Calendrier des livraisons
+- Tâches prioritaires
+- Messages importants
+
+#### Gestion des Commandes
+- Création de nouvelles commandes
+- Suivi des commandes en cours
+- Historique des commandes
+- Gestion des réclamations
+
+## 4. Delivery Person Dashboard
+
+### Sidebar Elements
+- 🚚 Mes Livraisons
+- 📍 Navigation
+- 📱 Contact Client
+- 📝 Rapports
+
+### Layouts Détaillés
+
+#### Mes Livraisons
+- Liste des livraisons du jour
+- Carte des itinéraires
+- Statut des livraisons
+- Historique des livraisons
+
+#### Navigation
+- Carte interactive
+- Optimisation des itinéraires
+- GPS intégré
+- Points de livraison
+
+## 5. Customer Service Dashboard
+
+### Sidebar Elements
+- 📞 Tickets Support
+- 👥 Base Clients
+- 📊 Statistiques
+- 📝 Rapports
+
+### Layouts Détaillés
+
+#### Tickets Support
+- Nouveaux tickets
+- Tickets en cours
+- Historique des résolutions
+- Performance du support
+
+## 6. Supervisor Dashboard
+
+### Sidebar Elements
+- 📊 Performance Équipe
+- 👥 Gestion Équipe
+- 📈 Rapports
+- ⚙️ Configuration
+
+### Layouts Détaillés
+
+#### Performance Équipe
+- KPIs par employé
+- Statistiques de productivité
+- Rapports de performance
+- Gestion des objectifs
+
+## Fonctionnalités Communes
+
+### Authentification et Sécurité
+- Connexion sécurisée
+- Détection automatique du type d'administrateur
+- Gestion des sessions
+- Logs d'activité
+
+### Interface Utilisateur
+- Design responsive
+- Thème personnalisable
+- Mode sombre/clair
+- Notifications en temps réel
+
+### Personnalisation
+- Widgets configurables
+- Favoris
+- Raccourcis personnalisés
+- Préférences utilisateur
