+++ frontend/web/admin/src/components/CustomerServiceDashboard/CustomerDatabase.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/CustomerDatabase.module.css';
+import Table from '../Table';
+import styles from '../style/CustomerDatabase.module.css';
 
 interface Customer {
   id: string;
