+++ frontend/web/admin/src/components/CustomerServiceDashboard/ServiceStatistics.tsx
@@ -1,25 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/StaffManagement.module.css';
+import Table from '../Table';
+import styles from '../style/ServiceStatistics.module.css';
 
-interface StaffMember {
+interface ServiceStatistic {
   id: string;
   name: string;
-  role: string;
+  value: number;
 }
 
-const StaffManagement: React.FC = () => {
-  const [staffMembers, setStaffMembers] = useState<StaffMember[]>([]);
+const ServiceStatistics: React.FC = () => {
+  const [serviceStatistics, setServiceStatistics] = useState<ServiceStatistic[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
-    const fetchStaffMembers = async () => {
+    const fetchServiceStatistics = async () => {
       setLoading(true);
       try {
-        const response = await axios.get('/api/staff-members');
-        setStaffMembers(response.data);
+        const response = await axios.get('/api/service-statistics');
+        setServiceStatistics(response.data);
       } catch (error) {
         if (error instanceof Error) {
           setError(error);
@@ -30,27 +30,27 @@ const StaffManagement: React.FC = () => {
         setLoading(false);
       }
     };
-    fetchStaffMembers();
+    fetchServiceStatistics();
   }, []);
 
   const columns = [
-    { key: 'id', label: 'Staff ID' },
+    { key: 'id', label: 'Statistic ID' },
     { key: 'name', label: 'Name' },
-    { key: 'role', label: 'Role' },
+    { key: 'value', label: 'Value' },
   ];
 
   return (
-    <div className={styles.staffManagementContainer}>
-      <h2>Staff Management</h2>
+    <div className={styles.serviceStatisticsContainer}>
+      <h2>Service Statistics</h2>
       {loading ? (
         <p>Loading...</p>
       ) : error ? (
         <p>Error: {error.message}</p>
       ) : (
-        <Table data={staffMembers} columns={columns} />
+        <Table data={serviceStatistics} columns={columns} />
       )}
     </div>
   );
 };
 
-export default StaffManagement;
+export default ServiceStatistics;
