+++ frontend/web/admin/src/components/CustomerServiceDashboard/SupportTickets.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/SupportTickets.module.css';
+import Table from '../Table';
+import styles from '../style/SupportTickets.module.css';
 
 interface SupportTicket {
   id: string;
