+++ frontend/web/admin/src/components/MasterSuperAdminDashboard/AdminActivityLog.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/AdminActivityLog.module.css';
+import Table from '../Table';
+import styles from '../style/AdminActivityLog.module.css';
 
 interface ActivityLogItem {
   timestamp: string;
