+++ frontend/web/admin/src/components/MasterSuperAdminDashboard/CompanyMetrics.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import styles from './style/CompanyMetrics.module.css';
+import styles from '../style/CompanyMetrics.module.css';
 
 interface RevenueData {
   totalRevenue: number;
