+++ frontend/web/admin/src/components/SecretaryDashboard/DailyOverview.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import styles from './style/DailyOverview.module.css';
+import styles from '../style/DailyOverview.module.css';
 
 interface DailyMetric {
   label: string;
