+++ frontend/web/admin/src/components/SecretaryDashboard/InvoiceManagement.tsx
@@ -1,25 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/ServiceReports.module.css';
+import Table from '../Table';
+import styles from '../style/InvoiceManagement.module.css';
 
-interface ServiceReport {
+interface Invoice {
   id: string;
-  name: string;
+  customerName: string;
   date: string;
 }
 
-const ServiceReports: React.FC = () => {
-  const [serviceReports, setServiceReports] = useState<ServiceReport[]>([]);
+const InvoiceManagement: React.FC = () => {
+  const [invoices, setInvoices] = useState<Invoice[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
-    const fetchServiceReports = async () => {
+    const fetchInvoices = async () => {
       setLoading(true);
       try {
-        const response = await axios.get('/api/service-reports');
-        setServiceReports(response.data);
+        const response = await axios.get('/api/invoices');
+        setInvoices(response.data);
       } catch (error) {
         if (error instanceof Error) {
           setError(error);
@@ -30,27 +30,27 @@ const ServiceReports: React.FC = () => {
         setLoading(false);
       }
     };
-    fetchServiceReports();
+    fetchInvoices();
   }, []);
 
   const columns = [
-    { key: 'id', label: 'Report ID' },
-    { key: 'name', label: 'Name' },
+    { key: 'id', label: 'Invoice ID' },
+    { key: 'customerName', label: 'Customer Name' },
     { key: 'date', label: 'Date' },
   ];
 
   return (
-    <div className={styles.serviceReportsContainer}>
-      <h2>Service Reports</h2>
+    <div className={styles.invoiceManagementContainer}>
+      <h2>Invoice Management</h2>
       {loading ? (
         <p>Loading...</p>
       ) : error ? (
         <p>Error: {error.message}</p>
       ) : (
-        <Table data={serviceReports} columns={columns} />
+        <Table data={invoices} columns={columns} />
       )}
     </div>
   );
 };
 
-export default ServiceReports;
+export default InvoiceManagement;
