+++ frontend/web/admin/src/components/SuperAdminDashboard/ReportsView.tsx
@@ -1,25 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/SupervisorReports.module.css';
+import Table from '../Table';
+import styles from './style/ReportsView.module.css';
 
-interface SupervisorReport {
+interface Report {
   id: string;
   name: string;
   date: string;
 }
 
-const SupervisorReports: React.FC = () => {
-  const [supervisorReports, setSupervisorReports] = useState<SupervisorReport[]>([]);
+const ReportsView: React.FC = () => {
+  const [reports, setReports] = useState<Report[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
-    const fetchSupervisorReports = async () => {
+    const fetchReports = async () => {
       setLoading(true);
       try {
-        const response = await axios.get('/api/supervisor-reports');
-        setSupervisorReports(response.data);
+        const response = await axios.get('/api/reports');
+        setReports(response.data);
       } catch (error) {
         if (error instanceof Error) {
           setError(error);
@@ -30,7 +30,7 @@ const SupervisorReports: React.FC = () => {
         setLoading(false);
       }
     };
-    fetchSupervisorReports();
+    fetchReports();
   }, []);
 
   const columns = [
@@ -40,17 +40,17 @@ const SupervisorReports: React.FC = () => {
   ];
 
   return (
-    <div className={styles.supervisorReportsContainer}>
-      <h2>Supervisor Reports</h2>
+    <div className={styles.reportsViewContainer}>
+      <h2>Reports View</h2>
       {loading ? (
         <p>Loading...</p>
       ) : error ? (
         <p>Error: {error.message}</p>
       ) : (
-        <Table data={supervisorReports} columns={columns} />
+        <Table data={reports} columns={columns} />
       )}
     </div>
   );
 };
 
-export default SupervisorReports;
+export default ReportsView;
