+++ frontend/web/admin/src/components/SuperAdminDashboard/StaffManagement.tsx
@@ -1,25 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Table from './Table';
-import styles from './style/ReportsView.module.css';
+import Table from '../Table';
+import styles from '../style/StaffManagement.module.css';
 
-interface Report {
+interface StaffMember {
   id: string;
   name: string;
-  date: string;
+  role: string;
 }
 
-const ReportsView: React.FC = () => {
-  const [reports, setReports] = useState<Report[]>([]);
+const StaffManagement: React.FC = () => {
+  const [staffMembers, setStaffMembers] = useState<StaffMember[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<Error | null>(null);
 
   useEffect(() => {
-    const fetchReports = async () => {
+    const fetchStaffMembers = async () => {
       setLoading(true);
       try {
-        const response = await axios.get('/api/reports');
-        setReports(response.data);
+        const response = await axios.get('/api/staff-members');
+        setStaffMembers(response.data);
       } catch (error) {
         if (error instanceof Error) {
           setError(error);
@@ -30,27 +30,27 @@ const ReportsView: React.FC = () => {
         setLoading(false);
       }
     };
-    fetchReports();
+    fetchStaffMembers();
   }, []);
 
   const columns = [
-    { key: 'id', label: 'Report ID' },
+    { key: 'id', label: 'Staff ID' },
     { key: 'name', label: 'Name' },
-    { key: 'date', label: 'Date' },
+    { key: 'role', label: 'Role' },
   ];
 
   return (
-    <div className={styles.reportsViewContainer}>
-      <h2>Reports View</h2>
+    <div className={styles.staffManagementContainer}>
+      <h2>Staff Management</h2>
       {loading ? (
         <p>Loading...</p>
       ) : error ? (
         <p>Error: {error.message}</p>
       ) : (
-        <Table data={reports} columns={columns} />
+        <Table data={staffMembers} columns={columns} />
       )}
     </div>
   );
 };
 
-export default ReportsView;
+export default StaffManagement;
