+++ frontend/web/admin/src/components/SupervisorDashboard/TeamPerformance.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './style/TeamPerformance.module.css';
+import styles from '../style/TeamPerformance.module.css';
 
 interface TeamPerformanceProps {
   teamName: string;
