+++ frontend/web/admin/src/layouts/SecretaryDashboard/SecretaryDashboard.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { DashboardLayout } from '../DashboardLayout/DashboardLayout';
-import DailyOverview from '../../components/DailyOverview'; // Corrected import
+import DailyOverview from '../../components/SecretaryDashboard/DailyOverview'; // Corrected import
 import OrderManagement from '../../components/OrderManagement'; // Corrected import
 import ClientManagement from '../../components/ClientManagement'; // Corrected import
-import InvoiceManagement from '../../components/InvoiceManagement'; // Corrected import
-import Schedule from '../../components/Schedule'; // Corrected import
+import InvoiceManagement from '../../components/SecretaryDashboard/InvoiceManagement'; // Corrected import
+import Schedule from '../../components/SecretaryDashboard/Schedule'; // Corrected import
 import styles from './style/SecretaryDashboard.module.css';
 
 export const SecretaryDashboard: React.FC = () => {
