+++ frontend/web/admin/src/layouts/SuperAdminDashboard/SuperAdminDashboard.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { DashboardLayout } from '../DashboardLayout/DashboardLayout';
-import StaffManagement from '../../components/StaffManagement'; // Corrected import
-import OrdersManagement from '../../components/OrdersManagement'; // Corrected import
+import StaffManagement from '../../components/SuperAdminDashboard/StaffManagement'; // Corrected import
+import OrdersManagement from '../../components/SuperAdminDashboard/OrdersManagement'; // Corrected import
 import CustomerManagement from '../../components/CustomerManagement'; // Corrected import
-import ReportsView from '../../components/ReportsView'; // Corrected import
-import Settings from '../../components/Settings'; // Corrected import
+import ReportsView from '../../components/SuperAdminDashboard/ReportsView'; // Corrected import
+import Settings from '../../components/SuperAdminDashboard/Settings'; // Corrected import
 import styles from './style/SuperAdminDashboard.module.css';
 
 export const SuperAdminDashboard: React.FC = () => {
